refactor(tapo, tapo-py): address the second round of schedule review feedback
Fixes a listing bug of my own making. `get_schedule_rules` counted parsed
rules against the device's `sum`, but skipped rules never reach `all`, so a
single unparseable rule meant the break could never fire: the loop re-sent the
same request until the page bound ran out, or returned `Err` if the firmware
rejected the out-of-range `start_index`. Raw rules fetched are now counted
against `sum`, and an empty page also ends the loop.

Restores the timer's loud failure. Sharing `DesiredStateRaw` made `on`
optional for the timer too, so a `get_countdown_rules` response with an empty
`desired_states` and no legacy `action` parsed as `None`, the rule was dropped,
and `get_timer` reported no timer while one was armed and about to fire. That
was a side effect, not a decision: `into_timer` is fallible again, so an
unreadable armed timer is an error, while an empty rule list still means no
timer. The schedule listing keeps its leniency, which is deliberate there.

Drops the speculative serde defaults on the listing response. `sum` and
`schedule_rule_max_count` are always present on a P110, so they are required
now and a genuinely broken response fails loudly instead of being masked — a
defaulted `sum` of 0 would have silently truncated the listing to one page.
`get_max_schedule_rules` returns the count directly as a result.

Uses the error keys the app itself uses: SCHEDULE_FULL, SCHEDULE_CONFLICT and
SCHEDULE_INDEX.

Splits the schedule request module into a folder, matching `set_device_info`:
`days_of_week`, `schedule_time`, `schedule_rule`, `raw` and `params`, each with
its own tests. The rule's fields are `pub(super)` so the wire conversion can
read them while staying private outside the module.

Also fixes an underflow in the examples, where a device reporting a maximum
below its current rule count would panic a debug build, and drops a no-op
carve-out from a test helper that would have silently turned a future
explicit-null case into a missing-key one.
