+++ tapo/src/responses/tapo_response.rs
@@ -59,17 +59,10 @@ fn error_kind(code: i64) -> &'static str {
         -1010 => "PUBLIC_KEY",
         -1501 => "LOGIN",
         9999 => "SESSION_TIMEOUT",
-        // Smart Devices: plug schedule rules.
-        //
-        // Unlike the codes above, these names describe behaviour observed on a
-        // P110 rather than reproducing names published by TP-Link, which does
-        // not document them and whose app carries the codes as bare integers.
-        // Adding a 33rd rule to a device reporting `schedule_rule_max_count:
-        // 32` returns -1802; adding a rule identical to an existing one
-        // returns -1803; editing an id the device does not hold returns -1805.
-        -1802 => "SCHEDULE_RULE_MAX_COUNT",
-        -1803 => "SCHEDULE_RULE_CONFLICT",
-        -1805 => "SCHEDULE_RULE_NOT_FOUND",
+        // Smart Devices: plug schedule rules
+        -1802 => "SCHEDULE_FULL",
+        -1803 => "SCHEDULE_CONFLICT",
+        -1805 => "SCHEDULE_INDEX",
         // SmartCam Devices
         -40106 => "UNSUPPORTED_METHOD",
         -40109 => "ONE_SECOND_REPEAT_REQUEST",
